% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xy_data.R
\name{get_xy_data}
\alias{get_xy_data}
\title{A function to generate a dummy x-y data set}
\usage{
get_xy_data()
}
\arguments{
\item{n}{number of points in the random scatter (default is 20000)}
}
\value{
tibble
}
\description{
This is a function to load a \code{n} + 500 by 2 data set, consisting of
an x and y column, into your R session.
}
