% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{xydata}
\alias{xydata}
\title{A dummy x-y data set}
\usage{
data(xydata)
}
\value{
Dataframe
}
\description{
This is a 25000 by 2 data set consisting of
an x and y column.
}
\details{
\describe{
\item{x.}{Random set of values}
\item{y.}{Random set of values}}
}
\keyword{datasets}
